Fixed hot reloading
